Rename files and folders to match Rescript convention
