+++ src/components/InfoRow/InfoRow.res
@@ -1,6 +1,6 @@
 // {..} means we are handling a JS object with an unknown
 // set of attributes
-@module external styles: {..} = "./infoRow.module.css"
+@module external styles: {..} = "./InfoRow.module.css"
 
 @react.component
 let make = (~leftComponent, ~rightComponent) => {
