+++ src/components/NavBar/NavBarWithLocationMenu.res
@@ -1,4 +1,4 @@
-@module external mapPinIcon: string = "./map-pin.svg"
+@module external mapPinIcon: string = "./mapPin.svg"
 
 @react.component
 let make = (~title, ~subtitle, ~onClick) => {
